sample data creating errors
This file will produce an objects array inside the application storing it in a redux state property named fileInputData. This state property is expecting to have data that is sorted by one of the columns in the csv file (proposal_id) and filtered for empty values. The lines of code doing this are in src/logic, in the function storeInput, lines 6 and 7. If you place a breakpoint there, the expected actions occur. If that breakpoitn is not hit and the application runs without the debug window open and without a breakpoint, the state property named fileInputData results to an empty array.